Moved some comments to docstrings in report designer.
Fixed a bug where some style properties set in subclasses were not getting
set. There is still some work to do with the initproperties: I don't want
to have to rely on putting some props in beforeInit() and others in afterInit():
they should all go in initProperties() and the framework can sort out which
need to be set before construction and which after. But this is a step.Moved some comments to docstrings in report designer.