Unified initProperties() in uiwx and dBizobj. The call order is: 	1) beforeInit() 	2) initProperties() 	3) afterInit()
Warning: this differs from how it was previously. It used to be that
initProperties() was called late in the init phase, and it is now called
very early (before the wx object has been created yet). If you find you
are getting Segmentation Faults or such, look to make sure you aren't doing
anything like addObject() or whatnot from within your initProperties() 
method.

While the initProperties() hook is called early in the init phase, the 
properties are actually set very late in the init phase (after afterInit()).

All property settings (Dabo-ized properties) should go into 
initProperties(), and you no longer need to know which properties 
need to be set at instantiation and which can be set later. When 
subclassing, always put the props you want to start out with in 
initProperties() and the framework will handle the details.

Of course, you can still send properties to the constructor as
usual.

Got rid of the initChildObjects() hook, because we do that differently
elsewhere (addControls()) and it doesn't appear to be used anywhere.

Made self.properties into a private self._properties. If we do want public
access to this (I can't see why we would), let's make it into a property
(self.Properties).

Put the abstract beforeInit(), afterInit(), initProperties() and initEvents()
hooks (and the _private counterparts) into dObject so they are available in all 
dabo objects. They'll only get called from objects that call them explicitly
though (dPemMixin and dBizobj).

Sorry for the length of this diff: lots of props had to be hit for this change
to work.