Updated the datanav classes to use the dGridX class instead of creating its own grid from wx classes.
Added many updates to the dGridX class to accomodate the workflow when creating a bizobj-driven grid. Also enhanced the buildFromDataSet() method so that you can pass along a list of columns from the data set *not* to include as columns; a dict containing field:order pairs to control column ordering; and a flag to turn off auto-sizing of the columns.
Added evt.stop() calls in the key processing code. Turns out that without them, pressing 'Enter' advanced the selected cell as well as triggered the onEnterAction() code.
Added the ability to defer the sorting to a custom method. This is necessary with bizobj-driven grids, since the sorting needs to be done in the cursor for it to 'stick'.
Added a 'CurrField' property that returns the bound Field for the currently selected column.

Added some methods to dPemMixin that deal with getting sizer information for the given control. This allows for introspection, so that a control can find out how it's being sized. This will be necessary for the Designer; it may have other applications, too.

Added a minor tweak to the layout() method of dPanel. It seems to have helped with some resizing issues.

Modified the parameters sent to the sizers. Now you can specify 'halign' and 'valign', instead of having to pass a single tuple for 'alignment'. You can use either way, but if alignment is passed, halign and valign are ignored.
Added Height and Width properties to sizers. These are read-only, and allow you to determine the current size of the sizer.