Added Stefano Masini's proof-of-concept changes to the report writer, after reading the code and agreeing with the design. My only concern with it is that it may not be as easy to dynamically generate report forms at runtime (the dicts within dicts made that really easy), but other than that I can see that your object design will be much more maintainable and extensible.
Stefano, I realize this was only a proof of concept, but if you could
take ownership of this section of Dabo and move this along, I can then
be freed up to work on the ui, demos, and documentation. I'll send you
a password giving you commit rights to the repository. Style guidelines
for Dabo include:

+ tabs, NOT spaces (don't argue, just accept it and move on <g>)
+ please try to explicitly name the imports (from module import Classdef,
  and not (from module import *)

I'm keeping my old reporting module as dabo/lib/reporting_old for now,
mostly because I'm in the middle of a project that uses it. When the 
new reporting library provides everything I need for my project, I'll
switch to the new reporting libary and ditch the old. I agree with your
comments on the frameset thing, but I needed to be able to specify a 
random number of columns and to use a paragraph flowable. Let's move 
forward and get it done correctly.

If you don't think you can commit time to implementing and maintaining 
the Dabo report writer, please let us know now, otherwise I'll expect
that you are working on this. Congratulations and welcome to Dabo! :)