Added dStatusBar class. As of now, it's just a minimal wrapper around th wx.StatusBar class. Changed dFormMixin to use this new class.
Added props/methods to dPemMixin to allow for the automatic drawing of borders. While this currentlt only has application in the Designer, it will be necessary to have these hooks when we get around to implementing drawing primatives.

Added a setAll() method to dPemMixin. It allows for recursive application of prop changes to children of a container. Full explanation and usage in the docstring.