Wrote work-around code in the dPageSelect class to handle the fact that the SetPageText() method, which is called when dPage.Caption is changed, is not implemented yet on OS X for wx.Choice controls.
Added a Pages property to the PageFrame classes. This is similar to Children, except it excludes the dropdown list and image list in the dPageList and dPageSelect classes.

Added a 'fontMetric() method to the UI. You can pass in several parameters, and it will return the (width, height) tuple for the specified text in the specified font. 

Set a minimum of one pixel for the minor dimension of a dLine. It was previously defaulting to zero, which made the line invisible.