Added self-test to the ListSorter class.
Cleaned up some of yesterday's changes to the datanav Page.py. Removed the SortOrderForm class, as this functionality already exists in Dabo. Removed some wx-specific code. Changed the logic so that the sort ordering form only appears when specifically requested.

Fixed a couple of nagging bugs in the SortingForm class, and added output to the self-test section. Changed the importing of the SortingForm class in the __init__.py.sortList() method to be more explicit.

Made _showStatusBar a class property of the dDialog class, since dialogs do not have status bars.

Removed 'dodefault()' call from dForm.

Added a flag to dPemMixin to prevent the firing of some events (LostFocus, Paint, Resize) if the control is being destroyed. Also added some wrapper calls to wx's ScreenToClient and ClientToScreen methods.

Added some protective code to calls to _createStatusBar() and getStatusBar() in dFormMixin.