Corrected the bug that happened whenever an item was deleted. I had been internally flagging an object as being in the process of being deleted (obj._finito) when it received a wx.Destroy() event. That was wrong; when an object is destroyed, that event is sent to the object as well as all of its containers. So when a page in a pageframe was removed, the pageframe and the form both got marked as being deleted, and would then stop responding to Paint and Size events. This bug was responsible for the erratic display behavior of items with grids, as grids add and delete columns continuously.
I also removed the 'lockScreen()' and 'unlockScreen()' methods from the form, and added them to each control, renamed as 'lockDisplay()' and 'unlockDisplay()'. This allows you to lock just the area of the control being updated if desired; you can still call self.Form.lockDisplay() if you need to lock the entire form.