Added a 'ChildBorn' event. This allows a container to know when a new child has been added to it. The child object reference is contained in evt.Child.
Modified the PageFrameMixin class so that append/insertPage takes a page class, not an instance. This allows None to be passed as the class, which results in the default PageClass value being used. Also added a 'SelectedPageNum' property that returns the index of the currently active page.

Datanav: cleaned up many wx-style method calls; converted them to dabo-style. Also changed the code that adds pages to the pageframe to pass classes instead of instances.

Updated the dGrid class to use the new lockDisplay() instead of lockScreen() calls. Added a call to lockDisplay() when auto-sizing a column, as there was a noticeable flicker when the number of rows was less than the full size of the grid.

Added the property 'ControllingSizer' to dPemMixin. This returns a reference to the sizer, if any, that controls that object's layout. Also wrapped several property defs with _().