Changed some of the encoding calls in dGrid to use wx.GetDefaultPyEncoding(), based on Robin Dunn's recommendation.
Further work on the DatePickerCtrl version of the dDateTextBox. There are still issues with binding to keystrokes: the control seems to have two separate focus states, and the 'inner' focus doesn't respond properly to keystrokes. For now, stick with the original dTextBox-based version (keep USE_ORIG_CLASS = True at the top of the file).