Made some mods to the text search function. I had been working with dEditor a lot lately, and some of the limitations of the FindDialog were bugging me. Pressing 'Enter' now fires the Find function, and reverse searches work with both the StyledTextCtrl and regular text fields. Search for 'Whole word' isn't implemented, yet, though.
Also added a dUtils script, which will serve to house all the quick little scripts that aren't specific to a particular class, in case they may prove useful elsewhere. The only function in it now is the 'reverseText()' function, which does just what it says: takes any string and returns the reverse of it.