Changed connParser to allow for a name element, instead of the user@host method. If the name element doesn't exist, user@host is still used.
Fixed dApp's loading of cnxml files to properly load them if the application
was started from outside of the application home directory. Previously, they
would only load if the main script was started from within the home directory.

Changed the loading of connection defs to load all *.cnxml as well as the
dbConnectionDefs.py if it exists. Previously, everything was mutually exlusive:
either default.cnxml was loaded OR *.cnxml was loaded. Only if no cnxml files 
were loaded would dbConnectionDefs be loaded.