dToolBar.py: We now have a basic Dabo-ized toolbar. It still needs some work, of course, but it's working!
dFormMixin.py: Added the code to handle toolbar creation, as well as a 
passthrough method to add buttons. Also made StatusBar and ToolBar regular 
properties, rather than having to call getters. Modeled the ToolBar creation 
code on the code for StatusBar, and added a ShowToolBar property to control 
whether the form creates a toolbar (default=False).

datanav/Form.py: This is the one place that used a toolbar, so that's been 
converted to use the new syntax.

propertyHelperMixin.py: added an optional 'defaultValue' to the 'extractKey()' 
method. The default is still None, so this won't affect any existing code.

uiwx/__init__.py: Added the method 'getBitmap()'. This wraps the 
wx.ArtProvider.GetBitmap() method, so that we can access the built-in 
graphics that comes with wx. We may need to resolve some of the names here, 
as they conflict with the dIcons names.

dSizerMixin.py: added some name wrappers for Horizontal and Vertical, so that 
we don't need to use the wx constants wx.HORIZONTAL and wx.VERTICAL when 
working with sizer flags.