Added the 'dToolForm' class to dForm.py. This is a form designed to look and act like a floating toolbar. Also modified the __init__ of dFormMixin so that if an explicit style is passed, it is respected instead of automatically setting the style to the wx default.
Changed the sizer mixin so that when you set the Sizer property, the old
sizer is removed. Turns out that this small change cured a lot of the seg
faults I was receiving under Linux. Also added a 'realease()' method to the
sizer mixin class.

Added support for toolbar labels to dToolBar. They don't appear under Gtk,
but they do under Windows.