Moved the serialization directory into reporting, as it is specific to reporting.
Note: I'm not sure I'm going to proceed with Stefano's implementation,
simply because I don't, in the end, completely grok it. Also, I'd be
backtracking in that I already had a basic designer working and will
otherwise need to reimplement it. I'll keep plugging away at it, but
may end up moving reporting to reporting_stefano, and reporting_old
back to reporting.