Continued reworking dColumn so that it can be useful as a real dObject. Set props programmatically or via user interface and they'll do the right thing. Currently this is only true with the column width, but it's a start. This is going to require lots more work to get right, but I hope to get lots of the stuff in dGrid.py refactored to be more encapsulated in the proper places. Everything to do with column properties should be defined inside the dColumn definition, for example. We are definitely on the right path with dGrid, but I think more sorting out what belongs in datanav.Grid versus what belongs in the base dGrid may be in order.
Fixed a problem with column order where it would go in order by string, 
ie: "10", "110", "120", ... , "20"...

Fixed column widths so that it really does get set in a reasonable 
precendence: 1: user settings, 2: field specs, 3: default or AutoSize.
Now, my test application saves and reloads the user settings for column
widths, pulls the width specified in fieldspecs for the others, and if
fieldspecs specifies "-1" for a column, that column gets auto sized.

Added a property to dColumn, GridColumnIndex, which figures out which
column, if any, the grid table has us mapped to.

I'm also going to commit a change to AppWizard that sets the column
widths to -1 by default (so that auto sizing of the columns is the
default, unless the user overrides).