Fixed some typos in the dbSQLite.py file. I should have caught them before, but I had to run out without testing.
Figured out what is preventing anything from being saved with SQLite data: the routine that determines the non-updatable fields depends on the description, and the current situation of determining the description is resulting in *all* fields being marked as non-updatable. We will have to use the code I added earlier today to create a backend-specific solution to this.