Added functions for determining the user's home directory path and dabo directory path.
The user's dabo directory is the subdirectory under the home directory
where user preference and setting information can be stored. On
Linux and Mac, this is something like:
	/home/pmcnett/.dabo.
On Windows, it is something like:
	c:\Documents and Settings\pmcnett\Application Data\Dabo

In one of my next commits, I'll be changing the file name of dUtils.py
to utils.py to fall in line with our naming conventions.