Fixed dConnectInfo to accept empty Port values.
Moved several methods, or parts of methods, out of dCursorMixin and into dBackend. These all contained behaviors that relied on the dbapi module returning an accurate description after a query, whether or not the query contained any records. SQLite doesn't follow that assumption, so these behaviors were moved into dBackend where they could be overridden for dbSQLite.

Fixed a tricky bug that would sometimes result in an incorrect RowNumber value being stored in a cursor.

Since SQLite databases are file-based, added support for POSIX-style '~' in paths for the database name. Corrected a bug that resulted in system tables being included in getTables() when they were supposed to be ignored. Also fixed some problems with the getFields() routines that would sometimes return the wrong information. And, of course, added SQLite-specific versions of the cursor methods that were moved into the dBackend class.