Modified dUserSettingProvider to save the settings file to the user's home directory instead of in all those .userSettings.ini files. You can now clean those buggers up and they'll no longer litter your hard drive.
Now, you'll find a .dabo directory in your home directory, or if you
are on Windows you'll find a Dabo directory in 
c:\Documents and Settings\<username>\Application Data. In this 
directory, you'll find one or more userSettings.ini files, with names
like userSettings-dabo.ini, or userSettings-recipedemo.ini. The 
specific part of the filename comes from app.getAppInfo("appShortName"),
which defaults to app.getAppInfo("appName").lower() if not set. If
appName isn't defined, that defaults to "Dabo".

Modified the getUserDaboDirectory() to return None if it wasn't able
to create the directory.

I also included code that would let application developers change the
name of the .dabo directory to anything they like.

In addition, I added code in the application init phase to set the 
caption of the main form to the application name. This only applies
if the mainform subclass doesn't already define a caption.

Following is a commit to daboide that exposes the setting of the app
name, so it is easy for individual applications to have their own names.
I'll commit a change to appRecipes that does this as well.