Slightly altered the layout of dOkayCancelDialog, and changed calls to Layout() to layout(). Took away the explicit captions for the buttons and set the ok button to the stock wx.ID_OK button, which draws a nice bitmap on the button like the cancel button has.
As far as laying out the buttons go, I think we should refactor it
to use wx.StdDialogButtonSizer, which automatically places these
buttons according to platform conventions. Here's the blurb from
the Changelog for wxPython 2.5.4.1 describing this:

"""
Added wxStdDialogButtonSizer, which is a a special sizer that knows
how to order and position standard buttons in order to conform to the
current platform's standards.  You simply need to add each `wx.Button`
to the sizer, and be sure to create the buttons using the standard
ID's.  Then call `Realize` and the sizer will take care of the rest.
"""

Some platforms (Mac, GTK) have the standard order of the buttons as
("Cancel", "OK"). Others (Win) tend to do ("OK", "Cancel"). Letting
this special sizer work out these details seems to be the smart way
to go.