Simplified dCursorMixin so that the getFields() function no longer requires you to pass a table name. If no name is passed, it uses its value of Table as a default.
Added a 'replaceFor' function to dBizobj. This allows you to replace the value in a field in the current record set with a new value. If no condition is passed, all records are updated. These updates happen only in the cursor; they aren't committed to the database until you call save().
Right now, you can only replace the field with a literal value; next I will add the ability to replace a field with a calculated value, such as the upper() of the field.