The problems with dDateTextBox still exist, so temporarily changed the datanav date editing control to a raw dTextBox, which don't exhibit the problems.
Made a minor changed to dControlItemMixin to route any unfound key
to a None key if it exists, otherwise an exception will be raised.
This lets me use a dDropdown as the control to select the foreign
key value. If the foreign key is NULL, I can have the dDropdown 
display "< None >".