Added new sizing options to dabo.ui.strToBmp. You can pass either a scaling ratio (where 1.0 is same size, 0.75 is 3/4 size, etc.), or absolute heights and widths.
Stored the hgap and vgap parameters to a dGridSizer as attributes, as there doesn't seem to be any way to determine them easily after the sizer is created.

Added properties to dBitmapButton that help control image sizing. They work OK, but I may still streamline some of the auto-resizing code.

Fixed a bug in dSizerMixin that reversed Width and Height values.