This fixes the problem with not having the correct field types defined before a query has been run when creating new records. If the type isn't available, simple skip all of the type coercion code, as it's pointless.
Problems may arise if the value returned from the control, which will almost always be a string, is incorrect for the underlying table. Next, I'll look into testing this with the different backend databases to see how each handles this situation. I will also look into the possibility of directly querying the database for type information, but this will not help when it comes to calculated columns.