Incorporated Adi's patch for dListControl, but enhanced it a bit. Instead of using the Activate event, which is really meant for Forms and Applications and not individual controls, I added a new ListSelection event. And, I bound Hit to the activation event, since that is the one that fires when the user presses enter or double-clicks.
I added new properties for MultipleSelect, HitIndex, and LastSelectedIndex.

Also, I removed the default onHit event callback to fall in line with the rest
of Dabo. But I added onHit and onListSelection callbacks to the test, and made
the test multiple-selectable.

Thanks Adi for your contribution, and let me know how my changes work for you.