Fixed a typo in dBizobj.py.
Changed the seek() method in dCursorMixin to actually move the record pointer. I don't know why it never did this before, as I can't find any code that acutally uses this command, but just in case someone relies on the pointer not moving, I added an optional parameter 'movePointer' that defaults to True. If you pass movePointer=False, it will return the result of the seek without moving the record pointer.

Both these catches are courtesy of Adi J. Sieker.