These changes add framework-level support for field-level validation.
All data controls have a property named 'FieldValidate' that is False by default. If it is set to True, the framework will then enforce validation for that field. I'm not sure if this is the best way to go; perhaps just ignoring non-changed fields will eliminate most of the potential wasted calls, and lack of code in the bizobj will handle the rest. But I put it in here for now as part of the testing; we can always remove it later if it proves unnecessary.

This is a preliminary commit, mostly to test on other platforms, and to get feedback from others. For developers, you simply need to create a method in your bizobj named 'fieldValidation()', and put your validation logic in it. If invalid values are found, simply return a non-emtpy string, and the value will not be propagated back to the DataSource.