Moved global settings importing to the very top of 'import dabo'.
dObject wasn't calling its superclass __init__ code. Fixed. dPemMixin
wasn't calling its superclass __init__code. Fixed.

Added new experimental setting "autoBindEvents" and default to False.
Added autoBindEvents() to eventMixin, called during the init phase
of the object. If the autoBindEvents setting is True, the code will
use introspection to determine if any class methods are of the on*
naming convention and, if so and if there is a corresponding event 
defined in, dabo.dEvents and automatic bindEvent() call will be 
generated.

To see this feature in action, see dabodemo/tutorial/autoBindEvents.py.