Added the new RegID property to dPemMixin. This is an optional property that can be used to simplify object referencing in complex forms.
I originally had a object retrieval method in the dFormMixin class, but the name resolution code seemed much better and cleaner that I removed it. Now, no matter where the object is in the containership hierarchy of the form, you can reference it using simple dot notation. E.g., if the object's RegID is 'cellphone', any code in the form can reference it as 'self.cellphone', and any other object on the form can reference it as 'self.Form.cellphone'.

Also cleaned up some code for the BorderStyle property.