Enhanced dColumn to maintain a GridCellAttr object which will be fed to grid.Table at the appropriate time. This GridCellAttr will be the central point to maintain the properties of the cells in the column.
Added two new properties to dColumn, HorizontalCellAlignment and
VerticalCellAlignment, which use the above method. Other new 
properties we can add include Font and Fore/BackColor.

Added an automatic right alignment when the datatype is numeric,
with the intent to refactor it later for this automatic behavior
to be optional, but the default.