Removed the infoLog for the kind != 0, as I think it is unimportant.
Moved the setting of the cell editors/renderers into dColumn, and they only
get set now when the grid requests the attr object from the table. IOW, there
should be some performance improvement with large grids.

New properties in dColumn are:
	CustomEditor: Set to override the default editor. Default None.
	CustomRenderer: Set to override the default renderer. Default None.
	Editor: Returns the editor class used for cells in the column. Read-only.
	Renderer: Returns the renderer class used for cells in the column. Read-only.
	ListEditorChoices: The list of choices that will appear in the list editor if applicable.

Converted grid.Columns into a true property - it was just an undocumented attribute
before.

Added the dabo.autoBindEvents=True to the test. Run dGrid.py and you'll see that 
autobinding is working on the checkbox, however also please note a problem:

Edit a grid cell and then move to a different cell. A dEvents.GridCellEdited
event gets fired (good), but because of dabo.autoBindEvents the dGrid.onGridCellEdited()
method is bound automatically. However, this isn't the intended implementation in
the dGrid code, which has _onGridCellEdited() receiving the event and then calling
onGridCellEdited() with additional arguments. IMO, we can just remove onGridCellEdited(),
and even _onGridCellEdited(), and instead put the helper code inside the event itself,
available in GridCellEdited.EventData. Does that sound right, Ed? Before turning 
auto event binding on globally, we need to check for other occurances of this problem.