Changed the default sortArrowColor to Silver, I think it looks better. I'll get that property-ized soon.
Commented out the SetColLabelAlignment call: appears to have no effect. I'll
get that propertyized in dColumn soon.

Removed getHTML(), as that's not used anywhere anymore and is very hokey.

Consolidated the __main__ test and the _dGrid_test class - the __main__ test
now uses the _dGrid_test class instead of defining a separate grid.