Added 3 new grid header events: 	GridHeaderMouseEnter 	GridHeaderMouseLeave 	GridHeaderMouseRightClick
Wrote a _persist() method for dColumn, that simplifies persisting column props
like Order and Width to the user settings table. Added an optional property
to autoSizeCol() to persist the data. If the call to autoSizeCol() was the 
result of the user asking for it, we persist the new size. Otherwise, we don't.

Made a right-click context menu for the grid header region, with similar hook
as provided in the right-click context menu for the grid. Moved the autoSizeCol
call from the MouseRightUp event into an option in the default context menu
instead. This will allow users to have multiple choices for right-clicks in
the header region. 

Moved a few functions around to be in alpha order within their section, but
didn't get very far with that.