This merges my temporary branch back into the trunk. All the changes are logged as per my commits over the past few days.
I was going to hold off on doing this until most of my work on dGrid was done,
but I left trunk's dGrid in a semi-unusable state and this new one is looking
very good, so what the hey. 

To summarize, dGrid is now better performing, is more flicker-free, has a
better API (more properties, better encapsulation of responsibilities,
better control over minute details while still being really easy to use),
and it fits in better with overall Dabo (props for "DataSource" at the Grid
level and "DataField" at the column level, which can be links to the bizobj
or to a "dataset", such as what you get from bizobj.getDataSet() or really
any tuple of dicts).