I'm programming an application and it has a radiogroup attached to a bizobj, and I have the Hit event of the radiogroup bound to a method that conditionally enables/disables other controls on the screen depending on the value of the underlying bizobj field.
The problem is the value doesn't get flushed to the bizobj until focus 
leaves the control. I went ahead and made the decision that Mac has this 
right: non-text controls should flush their value at the instant the value
changes. 

So... I changed dRadioGroup and dCheckBox to do this unconditionally for
all platforms. We'll have to review other controls to see what they do
(dListBox, etc.) and fix as necessary.

I also fixed a subtle bug in flushValue(), which wasn't getting to the 
code to set _oldVal back to None, resulting in some values not getting
flushed back to the underlying bizobj.