There were at least two problems contributing to the segfaults with the editors, and while I was partially successful in getting it working dynamically on an on-demand basis, that just caused other problems. So, I resorted to assigning the editor for the entire column when the cell gets selected, which achieves the goals.
With these changes I cannot get DesignerPropSheet to segfault, and
it appears to work just fine, although the props aren't getting saved
to the objects because when I change something on button1, and then
go to button2 and back, the original values are there again. I'm 
sure that is something with the PropSheet and not with dGrid, because
when I edit a grid value in the dGrid test, the change does make it
to the underlying dataset.

Tested on Linux. Will followup on Mac and Windows to be sure. Yerba
mate time.