Added the ability to reportWriter for frameset objects to figure out how much height they need based on what Platypus reports the height of the highest flowable is. This means I can print my customer invoices, with descriptions that can wrap to any length. Well, wrapping to the next page won't work yet, though.
Added showExpr tag to the report format, which is an expression to evaluate 
at runtime to determine if the object should get printed. This is analagous
to VFP's Print When. My invoices don't print my billing rate and number of
hours if the invoice line amount is fixed instead of calculated.

Added reportUtils function getTestCursorXmlFromDataSet, which returns xml
text you can copy/paste into a new .rfxml file for your test dataset.