Modified getPropertyList(), and added getMethodList() and getEventList(). These functions all return a sorted list of available PEMS for the object, from the point of view of the user. IOW, nothing starting with _ is included, and for properties and events, only uppercase first letters are included. For methods, only lowercase first letters are included.
TODO: make those methods private (_getPropertyList(), etc.)

Changed getPropertyInfo's "show/editInDesigner" keys to "readable" and
"writable". I have a small change to daboide to fix this change there as
well, and will commit next. Removed the editorInfo key as it is unused.

TODO: remove the getEditorInfo() functions from the various objects that
still have them, if any.

Added getApiDoc() function to dObject, which returns a single-page HTML
documentation of the object. For now, all it does is display a list of
PEM's, with bold-face on the items that are defined in this class, and
normal-face for those items defined in superclasses.

Fixed some typos in dEvents that became exposed during my testing of the 
above.

I'm not sure about the ultimate location for getApiDoc... perhaps it really
doesn't belong directly in dObject but in a separate documentation utility.
But it is a good start. Try it:

import dabo
dabo.ui.loadUI("wx")

testobj = dabo.ui.dGrid

html = testobj.getApiDoc()
f = open("/home/pmcnett/test.html", "w")
f.write(html)
f.close()