Added error handling to the continueEvent() and discontinueEvent() methods. Seems that occasionally the uiEvent object was a Dabo event, not a wxEvent.
If the event has a GetIndex() or GetSelection() method available, the EventData processing will get that value and store it in EventData["index"].

Changed key binding in dPemMixin to be case-insensitive. Previously, self.bindKey("Enter", self.someEvent) would have thrown an error.

Changed the dControlItemMixin class to support multiple-selection controls in its getPositionValue() method.

Moved the atts common to all list item controls to the dControlItemMixin class and out of the specific classes.

Began the conversion of event bindings in list item controls. dListBox is working fine; dListControl is pretty close, but still needs a bit of testing. I haven't touched the dropdown/combo classes yet except to remove the common att definitions.