Added the option of selecting the newly-added item to dControlItemMixin's append/insertItem methods.
Added an attribute 'appendOnEnter' to dComboBox (default=False). When True, pressing Enter after adding text to the combo adds the item to the control's Choices. There is a beforeAppendOnEnter() hook that allows the programmer to conditionally disallow the append, or to just modify the text to be entered (e.g., force to upper case). And, of course, a matching afterAppendOnEnter() hook method if any post-processing is needed.

I decided not to raise ListSelection events for this control, since there is no matching ListDeselection event, and it does seem a bit redundant for combo boxes.