Improved the paragraph object behavior in the report writer: it will now give spacing between individual paragraphs if blank lines exist in the source field.
Reworked the "all records" versus "just this record" choices in the datanav
print preview to be "list format" and "expanded format" instead. List format
makes a report that mimics the browse grid, and expanded format makes a report
that mimics the layout of the edit page. Both print all records in the dataset,
and the expanded format puts a pagebreak between records.