Added a utility method called 'dictStringify' to convert the keys of a dictionary from unicode values to regular string values. Normally this isn't a problem, except when you need to expand a dict to named arguments using the **dict construction.
Enhanced xmltodict so that you can pass an optional list of attributes to skip. There are some attributes stored in a Designer form that are needed for the Designer, but not needed when the form is created live. This avoids the need to constantly filter them out of the returned dict.