First MAJOR step forward in using the files created by the Designer to actually create live Dabo forms!
The dForm class now accepts an argument 'src', which can be a cdxml file, or the contents of such a file. If that argument is passed, the form will re-create the form as created in the Designer environment. Note that you must use the current version of the Designer to create the cdxml file, as there have been structural changes to make this possible.

The sizers have also been modified by moving some of the enhancements that were formerly just in the Designer subclasses into the main class, as these will be needed to re-create them from the cdxml file.