Fixed a couple spelling errors in dCheckBox...
Regarding 3-state, I guess it is okay how it is. Basically, if you go 3-state,
if I understand it correctly it isn't going to propagate the value to the
underlying bizobj. While I don't really like that, I also can't come up with
a use-case for a 3-state checkbox attached to a logical field, unless you want
True/False/None to propagate back. The real use for 3-state is when the value
represents more than one object, and there are mixed True/False values in those
objects.

Anyway, it looks fine for now, thanks for adding it.