Added storing of position for wx.ContextMenuEvent EventData.
Modified the showContextMenu() method of dPemMixin to use ScreenToClient() to get a position relative to the control.

Simplified the reference to dPage in dPageFrameMixin.

Added some of the dObject abilities to handle property-named parameters to dSizer. I originally tried having dSizerMixin descend from dObject, but that just seemed too messy.

Correct missing _baseClass assignment in dPage.  Simplified the _getPagePosition() method.