I was really tired last night when I reported that no mouse events were working for dButton. I used a print statement like 'print "right click"' but I omitted the print statement so it was just a string literal.
Anyway, for dButton all the mouse events are working, and I've updated 
the test to respond to MouseRightClick and ContextMenu. Ed, please run
dButton.py and confirm that when you right click, you get messages for
both events, but when you press your context menu key on your keyboard,
you only get ContextMenu.

AFAICS, everything is working as it should. However, perhaps things are
different on Mac, which I haven't tested on.