Added a tools directory to the top-level dabo directory (alongside setup.py), and added a script to the tools directory called reportMixedIndentation.py. This script scans the directory tree, starting from where the script was launched from, and reports on files and lines with mixed indentation, and then gives you the opportunity to hand-edit each line in vim.
The rest of the changes are a first pass at fixing the indentation problems,
done while testing the reportMixedIndentation script.

Another item that may end up in this tools directory would be a checkConfig.py,
which people could run to get a list of failed dependencies or suggested
libraries to install.